# site/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/busca_cpf":{"get":{"summary":"Busca","description":"Busca informações no Portal da Transparência com base em CPF ou CNPJ informado.\n\n### Parâmetros:\n- **query** (`str`, obrigatório): CPF, Nome ou NIS a ser pesquisado.\n- **force_max_results** (`bool`, opcional): Se `True`, força a busca por todos os resultados disponíveis. AVSISO: Coletar grandes quantidades de dados pode levar a bloqueios temporários. Padrão: `True`.\n- **extract_details** (`bool`, opcional): Se `True`, extrai detalhes adicionais dos resultados. Padrão: `False`.\n- **search_result_limit** (`int`): Número máximo de resultados a serem retornados. Padrão: `10`.\n- **servidor_publico** (`bool`, opcional): Filtra por servidor público. Padrão: `None`.\n- **beneficiario_programa_social** (`bool`, opcional): Filtra por beneficiários de programas sociais. Padrão: `None`.\n- **portador_cpgf** (`bool`, opcional): Filtra por portadores do CPGF. Padrão: `None`.\n- **portador_cpdc** (`bool`, opcional): Filtra por portadores do CPDC. Padrão: `None`.\n- **sancao_vigente** (`bool`, opcional): Filtra por pessoas com sanções vigentes. Padrão: `None`.\n- **ocupante_imovel_funcional** (`bool`, opcional): Filtra por ocupantes de imóvel funcional. Padrão: `None`.\n- **possui_contrato** (`bool`, opcional): Filtra por pessoas com contratos com o governo. Padrão: `None`.\n- **favorecido_recurso** (`bool`, opcional): Filtra por favorecidos com recursos públicos. Padrão: `None`.\n- **emitente_nfe** (`bool`, opcional): Filtra por emitentes de Nota Fiscal Eletrônica. Padrão: `None`.\n\n### Retorno:\n- `dict`: Um dicionário com os resultados encontrados no Portal da Transparência.\n\n### Exemplo de uso:\n```bash\ncurl -X GET \"http://localhost:8000/busca?query=12345678901&max_results=true&extract_details=false\"\n```","operationId":"busca_busca_cpf_get","parameters":[{"name":"query","in":"query","required":false,"schema":{"type":"string","title":"Query"}},{"name":"force_max_results","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":true,"title":"Force Max Results"}},{"name":"extract_details","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":false,"title":"Extract Details"}},{"name":"search_result_limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Search Result Limit"}},{"name":"servidor_publico","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":false,"title":"Servidor Publico"}},{"name":"beneficiario_programa_social","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Beneficiario Programa Social"}},{"name":"portador_cpgf","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Portador Cpgf"}},{"name":"portador_cpdc","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Portador Cpdc"}},{"name":"sancao_vigente","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Sancao Vigente"}},{"name":"ocupante_imovel_funcional","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Ocupante Imovel Funcional"}},{"name":"possui_contrato","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Possui Contrato"}},{"name":"favorecido_recurso","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Favorecido Recurso"}},{"name":"emitente_nfe","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Emitente Nfe"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/busca_cnpj":{"get":{"summary":"Busca Cnpj","description":"Busca informações no Portal da Transparência com base em CPF ou CNPJ informado.\n\n### Parâmetros:\n- **query** (`str`, obrigatório): CPF, Nome ou NIS a ser pesquisado.\n- **force_max_results** (`bool`, opcional): Se `True`, força a busca por todos os resultados disponíveis. AVSISO: Coletar grandes quantidades de dados pode levar a bloqueios temporários. Padrão: `True`.\n- **extract_details** (`bool`, opcional): Se `True`, extrai detalhes adicionais dos resultados. Padrão: `False`.\n- **search_result_limit** (`int`): Número máximo de resultados a serem retornados. Padrão: `10`.\n- **tipo_natureza_juridica** (`int`, opcional): Tipo de natureza jurídica da empresa. Padrão: `None`.\n- **uf_pessoa_juridica** (`str`, opcional): UF da pessoa jurídica da empresa no formato \"XX\". Ex: \"SP\", \"RJ\", \"MG\". Padrão: `None`.\n- **municipio** (`str`, opcional): Geocode do município da empresa, ex: \"3304557\" para o município do Rio de Janeiro. Veja: https://www.ibge.gov.br/explica/codigos-dos-municipios.php para mais detalhes. Padrão: `None`.\n- **valor_gastos_diretos_de** (`float`, opcional): Valor mínimo dos gastos diretos. Padrão: `None`.\n- **valor_gastos_diretos_ate** (`float`, opcional): Valor máximo dos gastos diretos. Padrão: `None`.\n- **valor_transferencia_de** (`float`, opcional): Valor mínimo da transferência. Padrão: `None`.\n- **valor_transferencia_ate** (`float`, opcional): Valor máximo da transferência. Padrão: `None`.\n- **sancao_vigente** (`bool`, opcional): Se o CNPJ está sob sanção vigente. Padrão: `None`.\n- **emitente_nfe** (`bool`, opcional): Se o CNPJ é emitente de NFE. Padrão: `None`.\n\n### Retorno:\n- `dict`: Um dicionário com os resultados encontrados no Portal da Transparência.\n\n### Exemplo de uso:\n```bash\ncurl -X GET \"http://localhost:8000/busca_cnpj?query=12345678000195&max_results=true&extract_details=false\"\n```","operationId":"busca_cnpj_busca_cnpj_get","parameters":[{"name":"query","in":"query","required":false,"schema":{"type":"string","title":"Query"}},{"name":"force_max_results","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":true,"title":"Force Max Results"}},{"name":"extract_details","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":false,"title":"Extract Details"}},{"name":"search_result_limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Search Result Limit"}},{"name":"tipo_natureza_juridica","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tipo Natureza Juridica"}},{"name":"uf_pessoa_juridica","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uf Pessoa Juridica"}},{"name":"municipio","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Municipio"}},{"name":"valor_gastos_diretos_de","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Valor Gastos Diretos De"}},{"name":"valor_gastos_diretos_ate","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Valor Gastos Diretos Ate"}},{"name":"valor_transferencia_de","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Valor Transferencia De"}},{"name":"valor_transferencia_ate","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Valor Transferencia Ate"}},{"name":"sancao_vigente","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Sancao Vigente"}},{"name":"emitente_nfe","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Emitente Nfe"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}